added require auth
